refactor: move code around to the right places
