+++ packages/analytics-js-common/src/utilities/errors.ts
@@ -1,7 +1,13 @@
 import { isTypeOfError } from './checks';
 import { stringifyWithoutCircular } from './json';
 
-const MANUAL_ERROR_IDENTIFIER = '[MANUAL ERROR]';
+const MANUAL_ERROR_IDENTIFIER = '[MANUALLY DISPATCHED ERROR]';
+
+const hasStack = (err: any) =>
+  !!err &&
+  (!!err.stack || !!err.stacktrace || !!err['opera#sourceloc']) &&
+  typeof (err.stack || err.stacktrace || err['opera#sourceloc']) === 'string' &&
+  err.stack !== `${err.name}: ${err.message}`;
 
 /**
  * Get mutated error with issue prepended to error message
@@ -20,10 +26,13 @@ const getMutatedError = (err: any, issue: string): Error => {
 };
 
 const dispatchErrorEvent = (error: any) => {
-  if (isTypeOfError(error)) {
-    error.stack = `${error.stack ?? ''}\n${MANUAL_ERROR_IDENTIFIER}`;
+  if (isTypeOfError(error) && hasStack(error)) {
+    let stack = error.stack ?? error.stacktrace ?? error['opera#sourceloc'] ?? '';
+    stack = `${stack}\n${MANUAL_ERROR_IDENTIFIER}`;
+    // eslint-disable-next-line no-param-reassign
+    error.stack = stack;
   }
   (globalThis as typeof window).dispatchEvent(new ErrorEvent('error', { error }));
 };
 
-export { getMutatedError, dispatchErrorEvent, MANUAL_ERROR_IDENTIFIER };
+export { getMutatedError, dispatchErrorEvent, MANUAL_ERROR_IDENTIFIER, hasStack };
